Break out queries into separate file
